Text Extractor App/requirements.txt
flask
opencv-python
Pillow
numpy
random2
strings
pytesseract
